BoxedWineUI: added 64-bit windows build.  Also added a very empty splash screen for Boxedwine.exe so that when the user launches a game from the UI they can tell something started.

+++ tools/ui/src/wxWidgets/wxUtils.cpp
@@ -9,7 +9,7 @@
 
 wxArrayString wxUtilRemoveDuplicates(const wxArrayString& values) {
     wxArrayString results; // new empty wxArrayString
-    for(int i=0; i<values.GetCount(); i++)
+    for(int i=0; i<(int)values.GetCount(); i++)
     {
         const wxString &s = values[i];
         if( results.Index(s) == wxNOT_FOUND )
